Package Refactoring
Created a new external package, for things that are not internal to the
simulation. Under that, created a new visual package, for things that
are either graphical or printed to the user.

Auto-refactored various imports to support the above.
