--- src/com/islands/games/lifesim/Informable.groovy
@@ -1,5 +1,0 @@
-package com.islands.games.lifesim
-
-interface Informable {
-    Map info()
-}
