+++ src/com/islands/games/lifesim/external/ConfigurationManager.groovy
@@ -1,4 +1,4 @@
-package com.islands.games.lifesim
+package com.islands.games.lifesim.external
 
 /**
  * Management class for configuration reads and writes.
