+++ src/com/islands/games/lifesim/external/Informable.groovy
@@ -1,0 +1,5 @@
+package com.islands.games.lifesim.external
+
+interface Informable {
+    Map info()
+}
