+++ src/com/islands/games/lifesim/external/SQLManager.groovy
@@ -1,4 +1,4 @@
-package com.islands.games.lifesim
+package com.islands.games.lifesim.external
 
 /**
  * Class to manage access to our SQL database.
