+++ src/com/islands/games/lifesim/external/visual/Printable.groovy
@@ -1,4 +1,6 @@
-package com.islands.games.lifesim
+package com.islands.games.lifesim.external.visual
+
+import com.islands.games.lifesim.external.Simulation
 
 trait Printable {
     static void print(String msg) {
