+++ src/com/islands/games/lifesim/life/Person.groovy
@@ -1,11 +1,11 @@
 package com.islands.games.lifesim.life
 
-import com.islands.games.lifesim.Printable
-import com.islands.games.lifesim.Simulation
-import com.islands.games.lifesim.Time
+import com.islands.games.lifesim.external.visual.Printable
+import com.islands.games.lifesim.external.Simulation
+import com.islands.games.lifesim.external.Time
 import com.islands.games.lifesim.society.Tribe
 
-import com.islands.games.lifesim.Random
+import com.islands.games.lifesim.external.Random
 
 /**
  * Class representing a person -- whether that means a human or demi-human or what-have-you.
