+++ src/com/islands/games/lifesim/society/Tribe.groovy
@@ -1,10 +1,10 @@
 package com.islands.games.lifesim.society
 
-import com.islands.games.lifesim.Informable
-import com.islands.games.lifesim.Location
+import com.islands.games.lifesim.external.Informable
+import com.islands.games.lifesim.external.Location
 import com.islands.games.lifesim.life.Person
-import com.islands.games.lifesim.Simulation
-import com.islands.games.lifesim.Time
+import com.islands.games.lifesim.external.Simulation
+import com.islands.games.lifesim.external.Time
 import com.islands.games.lifesim.metaphysics.Advancement
 import com.islands.games.lifesim.metaphysics.AdvancementManager
 import com.islands.games.lifesim.metaphysics.Affinity
